Move short-id generator to lib/ids
It's an isomorphic, feature-neutral utility (depends only on nanoid), so it
belongs in the shared lib/ layer next to format.ts / workspace-name.ts, not
among server/'s feature modules. Renamed to lib/ids.ts and updated the two
importers (registry, view-builders) to @/lib/ids.
